Feat: Modify ball bounce behavior
When landing on a paddle, depending on where the ball lands it bounces with a different angle. If it lands in the middle of the paddle, it'll bounce off of it almost straight up. If it lands further to the right of the paddle or its left (on the top surface, not the sides) it'll bounce with a much bigger angle.